Remove dispatcher arguments
Can be captured by RemoteTestUtil at startup.
